% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_ppasls.R
\name{ppasls}
\alias{ppasls}
\title{Penalized Principal Asymmetric Least Square Regression (P\eqn{^2}AR)}
\usage{
ppasls(
  x,
  y,
  H = 10,
  C = 1,
  lambda = 0.001,
  gamma = 3.7,
  penalty = "grSCAD",
  max.iter = 100
)
}
\arguments{
\item{x}{A numeric matrix of predictors (n x p), where n is the number of samples and p is the number of predictors.}

\item{y}{A numeric response vector of length n.}

\item{H}{Number of quantile levels (slices) used for asymmetric loss.}

\item{C}{Penalty parameter for the loss function.}

\item{lambda}{Regularization parameter controlling sparsity (default: SCAD).}

\item{gamma}{Hyperparameter for SCAD or MCP penalty (default: 3.7 for group SCAD).}

\item{penalty}{Type of group penalty. One of \code{"grSCAD"}, \code{"grLasso"}, or \code{"grMCP"}.}

\item{max.iter}{Maximum number of iterations for convergence.}
}
\value{
A list with the following components:
\describe{
\item{evalues}{Eigenvalues of the estimated Mn matrix.}
\item{evectors}{Eigenvectors (columns) corresponding to the estimated central subspace directions.}
\item{x}{Original input predictor matrix.}
}
}
\description{
This function implements sparse sufficient dimension reduction using penalized principal asymmetric least squares regression (P\eqn{^2}AR), supporting group SCAD, MCP, and Lasso penalties.
}
\details{
The function estimates a sparse basis for the central subspace by solving the penalized principal asymmetric least squares problem. The penalty can be chosen from group SCAD, group Lasso, or group MCP.

For further details on the underlying method, see Soale et al. (2022) and Shin et al. (2024).
}
\examples{
\dontrun{
  set.seed(1)
  n <- 100; p   <- 10
  B <- matrix(0, p, 2)
  B[1,1] <- B[2,2] <- 1
  x <- MASS::mvrnorm(n, rep(0, p), diag(1,p))
  y <- (x \%*\% B[,1]/(0.5 + (x \%*\% B[,2] + 1)^2)) + 0.2*rnorm(n, 0, 1)
  fit <- ppasls(x, y, H = 10, C = 0.5, lambda = 0.1, gamma = 3.7,
              penalty = "grSCAD", max.iter = 100)
  fit$evectors[,1:2]
}
}
\references{
Soale, B. B., Artemiou, A., & Li, B. (2022). Sufficient dimension reduction via principal asymmetric regression. \emph{Statistica Sinica}.
}
