% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_ppl2svm.R
\name{ppl2svm}
\alias{ppl2svm}
\title{Penalized Principal L2-Hinge Support Vector Machine for Sparse Sufficient Dimension Reduction}
\usage{
ppl2svm(
  x,
  y,
  H = 10,
  C = 1,
  lambda = 0.001,
  gamma = 3.7,
  penalty = "grSCAD",
  max.iter = 100
)
}
\arguments{
\item{x}{A numeric matrix of predictors (n x p), where n is the number of samples and p is the number of predictors.}

\item{y}{A numeric response vector of length n.}

\item{H}{The number of slicing or quantiles (default: 10).}

\item{C}{Regularization parameter (default: 1).}

\item{lambda}{Penalty parameter controlling sparsity (default: 0.001).}

\item{gamma}{Regularization parameter for SCAD/MCP penalty (default: 3.7).}

\item{penalty}{Type of penalty: \code{"grSCAD"} (default), \code{"grLasso"}, or \code{"grMCP"}.}

\item{max.iter}{Maximum number of iterations for the algorithm (default: 100).}
}
\value{
A list with the following components:
\item{evalues}{Eigenvalues of the estimated matrix spanning the central subspace.}
\item{evectors}{Corresponding eigenvectors (columns) for the estimated sufficient directions.}
\item{x}{The original predictor matrix.}
}
\description{
This function implements the penalized principal L2-hinge support vector machine (P\eqn{^2}L2M) approach
for sparse sufficient dimension reduction (SDR)
}
\details{
The method estimates the sparse basis of the central subspace by minimizing a penalized principal L2-hinge SVM objective.
The algorithm supports group penalties to encourage row-wise sparsity, and efficiently computes the solution even in high-dimensional settings.
}
\examples{
\dontrun{
  set.seed(1)
  n <- 100; p <- 10
  B <- matrix(0, p, 2)
  B[1,1] <- B[2,2] <- 1
  x <- MASS::mvrnorm(n, rep(0, p), diag(1, p))
  y <- (x \%*\% B[,1] / (0.5 + (x \%*\% B[,2] + 1)^2)) + 0.2 * rnorm(n)
  fit <- ppl2svm(x, y, H = 10, C = 1, lambda = 0.2, gamma = 3.7,
                  penalty = "grSCAD", max.iter = 100)
  fit$evectors[, 1:2]
}

}
