% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_ppwlssvm.R
\name{ppwlssvm}
\alias{ppwlssvm}
\title{Penalized Principal Weighted Least Squares SVM (P\eqn{^2}WLSM) for Sparse Sufficient Dimension Reduction}
\usage{
ppwlssvm(
  x,
  y,
  H = 10,
  C = 1,
  lambda,
  gamma = 3.7,
  penalty = "grSCAD",
  max.iter = 100
)
}
\arguments{
\item{x}{Numeric predictor matrix (n x p), where n is the sample size and p is the number of predictors.}

\item{y}{Binary response vector of length n, coded as \code{-1, 1}.}

\item{H}{Number of quantile slices for principal machine construction (default: 10).}

\item{C}{Regularization parameter.}

\item{lambda}{Penalty parameter for sparsity. If NULL, cross-validation or an information criterion will be used for selection.}

\item{gamma}{Regularization parameter for SCAD/MCP penalty (default: 3.7).}

\item{penalty}{Penalty type: \code{"grSCAD"}, \code{"grLasso"}, or \code{"grMCP"}.}

\item{max.iter}{Maximum number of iterations (default: 100).}
}
\value{
A list with components:
\item{evalues}{Eigenvalues of the estimated sufficient dimension reduction matrix.}
\item{evectors}{Eigenvectors spanning the SDR subspace.}
\item{x}{(Optional) Original predictor matrix.}
}
\description{
Implements the penalized principal weighted least squares support vector machine (P\eqn{^2}WLSM) for sparse sufficient dimension reduction (SDR)
}
\details{
This function fits a penalized principal machine using a weighted least squares SVM loss, suitable for binary responsess. It supports group penalties for structured sparsity and variable selection. Only use with binary responses (\code{-1, 1}).
}
\examples{
\dontrun{
  set.seed(1)
  n <- 1000; p <- 10
  x <- MASS::mvrnorm(n, rep(0, p), diag(1, p))
  y <- (x \%*\% B[,1] / (0.5 + (x \%*\% B[,2] + 1)^2)) + 0.2 * rnorm(n)
  y.binary <- sign(y)
  fit <- ppwlssvm(x, y.binary, H = 10, C = 1, lambda = 0.0003, gamma = 3.7,
                  penalty = "grSCAD", max.iter = 100)
  fit$evectors[, 1:2]
}

}
