% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ppm_tune}
\alias{print.ppm_tune}
\title{Print a penalized principal machine cross-validation result}
\usage{
\method{print}{ppm_tune}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ppm_tune"} returned by \code{\link[=ppm_tune]{ppm_tune()}}.}

\item{...}{Currently ignored.}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Print a penalized principal machine cross-validation result
}
\seealso{
\code{\link[=ppm_tune]{ppm_tune()}}
}
